refactor: Renamed function (app)
