+++ src/migration/app/steps/create-file-path-maps/app/map-component-templates.ts
