+++ src/migration/app/steps/create-file-path-maps/app/map-route-templates.ts
