+++ src/migration/app/steps/create-file-path-maps/tests/map-components.ts
