+++ src/migration/index.ts
@@ -1,3 +1,3 @@
+export * from './app/index.js';
 export * from './ember-addon/index.js';
-export * from './ember-app/index.js';
 export * from './ember-engine/index.js';
