+++ tests/migration/app/index/sass.test.ts
@@ -1,21 +1,21 @@
 import { assertFixture, loadFixture, test } from '@codemod-utils/tests';
 
-import { migrateEmberApp } from '../../../../src/migration/ember-app/index.js';
+import { migrateApp } from '../../../../src/migration/app/index.js';
 import {
   inputProject,
   outputProject,
 } from '../../../fixtures/ember-app/sass/index.js';
 import { codemodOptions } from '../../../helpers/shared-test-setups/ember-app/sass.js';
 
-test('migration | ember-app | index > sass', function () {
+test('migration | app | index > sass', function () {
   loadFixture(inputProject, codemodOptions);
 
-  migrateEmberApp(codemodOptions);
+  migrateApp(codemodOptions);
 
   assertFixture(outputProject, codemodOptions);
 
   // Check idempotence
-  migrateEmberApp(codemodOptions);
+  migrateApp(codemodOptions);
 
   assertFixture(outputProject, codemodOptions);
 });
