+++ tests/migration/app/steps/create-options/javascript.test.ts
@@ -1,13 +1,13 @@
 import { assert, loadFixture, test } from '@codemod-utils/tests';
 
-import { createOptions } from '../../../../../src/migration/ember-app/steps/index.js';
+import { createOptions } from '../../../../../src/migration/app/steps/index.js';
 import { inputProject } from '../../../../fixtures/ember-app/javascript/index.js';
 import {
   codemodOptions,
   options,
 } from '../../../../helpers/shared-test-setups/ember-app/javascript.js';
 
-test('migration | ember-app | steps | create-options > javascript', function () {
+test('migration | app | steps | create-options > javascript', function () {
   loadFixture(inputProject, codemodOptions);
 
   assert.deepStrictEqual(createOptions(codemodOptions), options);
